Rewrote and cleaned up some code, still needs testing
