+++ src/server.cpp
@@ -1,5 +1,5 @@
-#include "../include/webserver.hpp"
-#include "../include/socket.hpp"
+#include "webserver.hpp"
+#include "socket.hpp"
 #include <cstdio>
 #include <cstring>
 #include <arpa/inet.h>
@@ -8,13 +8,17 @@
 #include <sys/event.h>
 #include <fstream>
 
-#define NUSERS 10
+// #define NUSERS 10
 
-struct uc { //this should go in the class!!
-    int uc_fd;
-    char *uc_addr;
-} users[NUSERS];
+// struct uc { //this should go in the class!!
+//     int uc_fd;
+//     char *uc_addr;
+// } users[NUSERS];
 
+
+#include "Request.hpp"
+#include "Response.hpp"
+#include <vector>
 
 bool	write_exit(std::string error)
 {
@@ -22,155 +26,77 @@ bool	write_exit(std::string error)
 	return (false);
 }
 
-/* find the index of a file descriptor or a new slot if fd=0 */
-int	Socket::connIndex(int fd) 
-{
-    int uidx;
-    for (uidx = 0; uidx < NUSERS; uidx++)
-        if (users[uidx].uc_fd == fd)
-            return uidx;
-    return -1;
-}
-
-int Socket::connAdd(int fd) 
-{
-    int uidx;
-    if (fd < 1) return -1;
-    if ((uidx = connIndex(0)) == -1)
-        return -1;
-    if (uidx == NUSERS) {
-        close(fd);
-        return -1;
-    }
-    users[uidx].uc_fd = fd; /* users file descriptor */
-    users[uidx].uc_addr = 0; /* user IP address */
-    return 0;
-}
-
-/* remove a connection and close its fd */
-int Socket::connDelete(int fd) {
-    int uidx;
-    if (fd < 1) 
-		return -1;
-    if ((uidx = connIndex(fd)) == -1)
-        return -1;
-
-    users[uidx].uc_fd = 0;
-    users[uidx].uc_addr = NULL;
-
-    /* free(users[uidx].uc_addr); */
-    return close(fd);
-}
-
-// void	Socket::watchLoop()
+// /* find the index of a file descriptor or a new slot if fd=0 */
+// int	Socket::connIndex(int fd) 
 // {
-	// struct kevent evList[32];
-	// int nev, i;
-	// struct sockaddr_storage addr;
-	// socklen_t socklen = sizeof(addr);
-	// int fd;
-	// while (1)
-	// {
-	// 	nev = kevent(kq, NULL, 0, evList, 32, NULL);
-	// 	if (nev < 1)
-	// 	{
-	// 		write_exit("kevent error");
-	// 		return ;
-	// 	}
-	// 	for (i = 0; i<nev; i++)
-	// 	{
-	// 		if (evList[i].flags & EV_EOF)
-	// 		{
-	// 			printf("Disconnect\n");
-	// 			fd = evList[i].ident;
-	// 			EV_SET(&evSet, fd, EVFILT_READ, EV_DELETE, 0, 0, NULL);
-	// 			if (kevent(kq, &evSet, 1, NULL, 0, NULL) == -1)
-	// 			{
-	// 				write_exit("kevent error");
-	// 				return ;
-	// 			}
-	// 			connDelete(fd);
-	// 		}
-	// 		else if ((int)evList[i].ident == listenfd)
-	// 		{
-	// 			printf("Here1\n");
-
-	// 			fd = accept(evList[i].ident, (struct sockaddr *)&addr, &socklen);
-	// 			if (fd == -1)
-	// 			{
-	// 				write_exit("accept error");
-	// 				return ;
-	// 			}
-	// 			if (connAdd(fd) == 0)
-	// 			{
-	// 				EV_SET(&evSet, fd, EVFILT_READ, EV_ADD, 0, 0, NULL);
-	// 				if (kevent(kq, &evSet, 1, NULL, 0, NULL) == -1)
-	// 				{
-	// 					write_exit("accept error");
-	// 					return ;
-	// 				}
-	// 				uint8_t buff[MAXLINE + 1];
-	// 				//snprintf((char*)buff, sizeof(buff), "HTTP/1.0 200 OK \r\nContent-Type: text/html\r\nContent-Length: 20\r\n\r\nWe socket thisssssss"); //can write formatted output to sized buf
-	// 				std::string   fileBuf;
-	// 				std::string line;
-	// 				std::ifstream   htmlFile;
-	// 				htmlFile.open("data/index.html");
-	// 				while (std::getline (htmlFile, line))
-	// 					fileBuf += line;
-	// 				snprintf((char*)buff, sizeof(buff), "HTTP/1.0 200 OK \r\nContent-Type: text/html\r\nContent-Length: %lu\r\n\r\n%s", fileBuf.length(), fileBuf.c_str());
-	// 				write(fd, (char*)buff, std::strlen((char*)buff));
-	// 				htmlFile.close();
-	// 			}
-	// 			else
-	// 			{
-	// 				printf("Connection refused\n");
-	// 				close(fd);
-	// 			}
-	// 		}
-	// 		else if (evList[i].filter == EVFILT_READ)
-	// 		{
-	// 			char buf[256];
-	// 			size_t bytes_read;
-
-	// 			bytes_read = recv(evList[i].ident, buf, sizeof(buf), 0);
-	// 			if ((int)bytes_read < 0)
-	// 				printf("%d bytes read\n", (int)bytes_read);
-	// 		}
-	// 	}
-	// }
+//     int uidx;
+//     for (uidx = 0; uidx < NUSERS; uidx++)
+//         if (users[uidx].uc_fd == fd)
+//             return uidx;
+//     return -1;
 // }
 
-bool	Socket::setUpConn(int kq, struct kevent evSet)
-{
-	if ((listenfd = socket(AF_INET, SOCK_STREAM, 0)) < 0) //AF_INET = internet socket, SOCK_STREAM = tcp stream
-		return (write_exit("socket error"));
-	int reuse; //this and setsockopt avoids the bind error and allows to reuse the address
-	if (setsockopt(listenfd, SOL_SOCKET, SO_REUSEADDR, (char *)&reuse, sizeof(int)) == -1)
- 		return(write_exit("reuse port error"));
-	//setting up address you're listening on
-	std::memset(&servAddr, '\0', sizeof(servAddr));
-	servAddr.sin_family		= AF_INET;
-	servAddr.sin_addr.s_addr = htonl(INADDR_ANY); // will respond to anything
-	printf("port: %d\n", this->port);
-	servAddr.sin_port		= htons(/*SERVER_PORT*/port); //port you're listening on
-	if ((bind(listenfd, (SA *) &servAddr, sizeof(servAddr))) < 0)//bind listening socket to address
-		return (write_exit("bind error"));
-	if ((listen(listenfd, 10)) < 0)
-		return (write_exit("listen error"));
-	// kq = kqueue();
-	EV_SET(&evSet, listenfd, EVFILT_READ, EV_ADD, 0, 0, NULL);//EV_SET is a macro that fills the kevent struct
-	if (kevent(kq, &evSet, 1, NULL, 0, NULL) == -1)
-		return (write_exit("kqueue/kevent error"));
-	// watchLoop();
-	return (true);	
+// int Socket::connAdd(int fd) 
+// {
+//     int uidx;
+//     if (fd < 1) return -1;
+//     if ((uidx = connIndex(0)) == -1)
+//         return -1;
+//     if (uidx == NUSERS) {
+//         close(fd);
+//         return -1;
+//     }
+//     users[uidx].uc_fd = fd; /* users file descriptor */
+//     users[uidx].uc_addr = 0; /* user IP address */
+//     return 0;
+// }
 
-}
+// /* remove a connection and close its fd */
+// int Socket::connDelete(int fd) {
+//     int uidx;
+//     if (fd < 1) 
+// 		return -1;
+//     if ((uidx = connIndex(fd)) == -1)
+//         return -1;
+
+//     users[uidx].uc_fd = 0;
+//     users[uidx].uc_addr = NULL;
+
+//     /* free(users[uidx].uc_addr); */
+//     return close(fd);
+// }
+
+// bool	Socket::setUpConn(int kq, struct kevent evSet)
+// {
+// 	if ((listenfd = socket(AF_INET, SOCK_STREAM, 0)) < 0) //AF_INET = internet socket, SOCK_STREAM = tcp stream
+// 		return (write_exit("socket error"));
+// 	int reuse; //this and setsockopt avoids the bind error and allows to reuse the address
+// 	if (setsockopt(listenfd, SOL_SOCKET, SO_REUSEADDR, (char *)&reuse, sizeof(int)) == -1)
+//  		return(write_exit("reuse port error"));
+// 	//setting up address you're listening on
+// 	std::memset(&servAddr, '\0', sizeof(servAddr));
+// 	servAddr.sin_family		= AF_INET;
+// 	servAddr.sin_addr.s_addr = htonl(INADDR_ANY); // will respond to anything
+// 	printf("port: %d\n", this->port);
+// 	servAddr.sin_port		= htons(/*SERVER_PORT*/port); //port you're listening on
+// 	if ((bind(listenfd, (SA *) &servAddr, sizeof(servAddr))) < 0)//bind listening socket to address
+// 		return (write_exit("bind error"));
+// 	if ((listen(listenfd, 10)) < 0)
+// 		return (write_exit("listen error"));
+// 	// kq = kqueue();
+// 	EV_SET(&evSet, listenfd, EVFILT_READ, EV_ADD, 0, 0, NULL);//EV_SET is a macro that fills the kevent struct
+// 	if (kevent(kq, &evSet, 1, NULL, 0, NULL) == -1)
+// 		return (write_exit("kqueue/kevent error"));
+// 	// watchLoop();
+// 	return (true);	
+
+// }
 
 int		Webserver::comparefd(std::vector<Socket> sckts, int eventfd)
 {
 	for (size_t i = 0; i < sckts.size(); i++)
 	{
-		if (sckts.at(i).listenfd == eventfd)
+		if (sckts.at(i).getListenfd() == eventfd)
 			return (1);
 	}
 	return (0);
@@ -183,8 +109,13 @@ void	Webserver::start(std::vector<Server> servers)
 		return ;//(false);
 	for (size_t i = 0; i < servers.size(); i++)
 	{
-		Socket sock(servers.at(i).getPort(0));
-		sckts.push_back(sock);
+		for (size_t j = 0; j < servers.at(i).getListens().size(); j++)
+		{
+			// servers.at(i).getListens.at(j).second;
+			// Socket sock(servers.at(i).getPort(0));
+			Socket sock(servers.at(i).getListens().at(j).second);
+			sckts.push_back(sock);
+		}
 	}
 	int kq = kqueue();
 	struct kevent evSet;
@@ -198,6 +129,15 @@ void	Webserver::start(std::vector<Server> servers)
 	struct sockaddr_storage addr;
 	socklen_t socklen = sizeof(addr);
 	int fd;
+	Request		*newReq;
+	Response	*newResp;
+	// std::vector<Location>	locations;
+	// uint8_t		*response; // needs to be malloced 
+
+	
+
+	// createLocation(locations);
+
 	while (1)
 	{
 		nev = kevent(kq, NULL, 0, evList, 2, NULL);
@@ -223,7 +163,7 @@ void	Webserver::start(std::vector<Server> servers)
 			}
 			else if (comparefd(sckts, (int)evList[i].ident) == 1)//(int)evList[i].ident == sckt1.listenfd || (int)evList[i].ident == sckt2.listenfd)
 			{
-				printf("Here1\n");
+				// printf("Here1\n");
 
 				fd = accept(evList[i].ident, (struct sockaddr *)&addr, &socklen);
 				if (fd == -1)
@@ -239,23 +179,31 @@ void	Webserver::start(std::vector<Server> servers)
 						write_exit("accept error");
 						return ;
 					}
-					uint8_t buff[MAXLINE + 1];
-					//snprintf((char*)buff, sizeof(buff), "HTTP/1.0 200 OK \r\nContent-Type: text/html\r\nContent-Length: 20\r\n\r\nWe socket thisssssss"); //can write formatted output to sized buf
-					std::string   fileBuf;
-					std::string line;
-					std::ifstream   htmlFile;
-					htmlFile.open("data/www/index.html");
-					while (std::getline (htmlFile, line))
-						fileBuf += line;
-					snprintf((char*)buff, sizeof(buff), "HTTP/1.0 200 OK \r\nContent-Type: text/html\r\nContent-Length: %lu\r\n\r\n%s", fileBuf.length(), fileBuf.c_str());
-					write(fd, (char*)buff, std::strlen((char*)buff));
-					htmlFile.close();
-				// }
-				// else
-				// {
-				// 	printf("Connection refused\n");
-				// 	close(fd);
-				// }
+					newReq = new Request(fd);
+					newReq->processReq();
+					newReq->printRequest();
+			// determine which server should handle this request (Request::identifyServer)
+			// 1. parse "listen" directives, if multiple matches with equal specificity:
+			// 2. parse "server name" directives find the server that corresponds to the request field's Host
+			// otherwise give it to the default one
+			
+					newResp = new Response(*newReq);
+					delete newReq;
+					newResp->prepareResponseGET(servers.at(0).getLocations()); // argument is ref to the Server
+					
+					// if (response)
+					// {
+					// 	// std::cout << "About to return " << newResp->getMsgLength() << "bytes: " << response << std::endl;
+					// 	send(fd, (char*)response, newResp->getMsgLength(), 0);
+					// 	delete response;
+					// }
+					delete newResp;
+			// }
+			// else
+			// {
+			// 	printf("Connection refused\n");
+			// }
+			// close(fd);
 			}
 			else if (evList[i].filter == EVFILT_READ)
 			{
@@ -268,7 +216,5 @@ void	Webserver::start(std::vector<Server> servers)
 			}
 		}
 	}
-
-
-	// return(sckt.setUpConn());
 }
+
