+++ src/socket.cpp
@@ -1,6 +1,84 @@
 #include "../include/socket.hpp"
 
+#define NUSERS 10
+
+struct uc { //this should go in the class!!
+    int uc_fd;
+    char *uc_addr;
+} users[NUSERS];
+
 Socket::Socket(unsigned short newport) : port(newport)
 {
+
+}
+
+int Socket::getListenfd()
+{
+    return(this->listenfd);
+}
+
+/* find the index of a file descriptor or a new slot if fd=0 */
+int	Socket::connIndex(int fd) 
+{
+    int uidx;
+    for (uidx = 0; uidx < NUSERS; uidx++)
+        if (users[uidx].uc_fd == fd)
+            return uidx;
+    return -1;
+}
+
+int Socket::connAdd(int fd) 
+{
+    int uidx;
+    if (fd < 1) return -1;
+    if ((uidx = connIndex(0)) == -1)
+        return -1;
+    if (uidx == NUSERS) {
+        close(fd);
+        return -1;
+    }
+    users[uidx].uc_fd = fd; /* users file descriptor */
+    users[uidx].uc_addr = 0; /* user IP address */
+    return 0;
+}
+
+/* remove a connection and close its fd */
+int Socket::connDelete(int fd) {
+    int uidx;
+    if (fd < 1) 
+		return -1;
+    if ((uidx = connIndex(fd)) == -1)
+        return -1;
+
+    users[uidx].uc_fd = 0;
+    users[uidx].uc_addr = NULL;
+
+    /* free(users[uidx].uc_addr); */
+    return close(fd);
+}
+
+bool	Socket::setUpConn(int kq, struct kevent evSet)
+{
+	if ((listenfd = socket(AF_INET, SOCK_STREAM, 0)) < 0) //AF_INET = internet socket, SOCK_STREAM = tcp stream
+		return (write_exit("socket error"));
+	int reuse; //this and setsockopt avoids the bind error and allows to reuse the address
+	if (setsockopt(listenfd, SOL_SOCKET, SO_REUSEADDR, (char *)&reuse, sizeof(int)) == -1)
+ 		return(write_exit("reuse port error"));
+	//setting up address you're listening on
+	std::memset(&servAddr, '\0', sizeof(servAddr));
+	servAddr.sin_family		= AF_INET;
+	servAddr.sin_addr.s_addr = htonl(INADDR_ANY); // will respond to anything
+	printf("port: %d\n", this->port);
+	servAddr.sin_port		= htons(/*SERVER_PORT*/port); //port you're listening on
+	if ((bind(listenfd, (SA *) &servAddr, sizeof(servAddr))) < 0)//bind listening socket to address
+		return (write_exit("bind error"));
+	if ((listen(listenfd, 10)) < 0)
+		return (write_exit("listen error"));
+	// kq = kqueue();
+	EV_SET(&evSet, listenfd, EVFILT_READ, EV_ADD, 0, 0, NULL);//EV_SET is a macro that fills the kevent struct
+	if (kevent(kq, &evSet, 1, NULL, 0, NULL) == -1)
+		return (write_exit("kqueue/kevent error"));
+	// watchLoop();
+	return (true);	
 
 }
